Fixed migrations to use correct type for foreign keys

+++ database/migrations/2019_08_10_164233_create_characters_table.php
@@ -15,7 +15,7 @@ class CreateCharactersTable extends Migration
     {
         Schema::create('characters', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('userId')->unsigned();
+            $table->bigInteger('userId')->unsigned();
             $table->foreign('userId')->references('id')->on('users');
             $table->string('name', 45);
             $table->string('player', 45);
