+++ database/migrations/2019_08_10_165541_create_abilities_table.php
@@ -15,7 +15,7 @@ class CreateAbilitiesTable extends Migration
     {
         Schema::create('abilities', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('characterId')->unsigned();
+            $table->bigInteger('characterId')->unsigned();
             $table->foreign('characterId')->references('id')->on('characters');
             $table->tinyInteger('archery');
             $table->tinyInteger('athletics');
